Remove CSV fetching
Previously I fetched CSV and JSON formats from the API, but with the current API
it's easier to just use JSON and then handle the formatting client side.

This is because I'm already deriving records client side in order to support incremental
updates. This is the most complicated part of the code so I don't want two implementations
of it.